Add files via upload
web application built to demonstrate various front-end and back-end development concepts. It showcases interactive features, a responsive design, and a user-friendly interface, with a focus on optimizing the user experience. The project utilizes modern web technologies, including HTML, CSS, JavaScript, and possibly some back-end logic, depending on the scope. The application is deployed on Netlify for easy access and live demonstrations.